requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioredis==1.3.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anthropic==0.59.0
anyio==4.9.0
argcomplete==3.6.2
asgi-csrf==0.11
asgi-lifespan==2.1.0
asgiref==3.9.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.39.14
botocore==1.39.14
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-default-group==1.2.4
cohere==5.16.1
colorama==0.4.6
colorlog==6.9.0
cryptography==45.0.5
datasette==0.65.1
distro==1.9.0
Django==5.2.4
djangorestframework==3.16.0
dnspython==2.7.0
drf-pydantic==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
eval_type_backport==0.2.2
Faker==37.4.2
fastapi==0.116.1
fastapi-cache==0.1.0
fastavro==1.11.1
filelock==3.18.0
flexcache==0.3
flexparser==0.4
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.3
greenlet==3.2.3
griffe==1.8.0
groq==0.30.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.5
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.1
hupper==1.12.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
json-schema-to-pydantic==0.2.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
logfire-api==4.0.0
loguru==0.7.3
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.12.2
mdurl==0.1.2
mergedeep==1.3.4
mistralai==1.9.3
multidict==6.6.3
nest-asyncio==1.6.0
openai==1.97.1
opentelemetry-api==1.35.0
packaging==25.0
passlib==1.7.4
peewee==3.18.2
platformdirs==4.3.8
pluggy==1.6.0
polyfactory==2.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==5.9.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-ai==0.0.43
pydantic-ai-slim==0.0.43
pydantic-graph==0.0.43
pydantic-settings==2.10.1
pydantic2==1.1.15
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
questionary==2.1.0
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
semver==3.0.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sqlparse==0.5.3
sse-starlette==2.4.1
starlette==0.47.2
tokenizers==0.21.2
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
yarl==1.20.1
zipp==3.23.0
